test(statusbar): 校正自定义主题夹具与 AppBar 计数类型 / Correct custom theme fixture and AppBar count type
中文：
- 将状态栏往返夹具的外观标记设为自定义，并使用非默认颜色，避免把预设标记与持久化的自定义标记作错误比较。
- 将 ABM_REMOVE 转为 DWORD，清除 AppBar 测试中 std::count 触发的 C4389。
English:
- Mark the status bar roundtrip appearance as custom and use a non-default color, avoiding an invalid comparison between preset and persisted custom markers.
- Cast ABM_REMOVE to DWORD to remove C4389 from std::count in the AppBar test.
验证 / Validation:
- scripts/test.bat name "^(general_settings|dock_and_window_rules)$": 2/2 passed, exit 0, CTest 7.07s, no compiler/linker warnings; .codex-probes/statusbar-implementation/02-tests-retry.log; test-run-e9c9bbd179b84ecea4f048a5dd50043b.xml.
- 前轮其余七项通过，相关输入未改，继续引用 / Reuse the other seven passing selections from the preceding run; their inputs are unchanged: localization_contract, settings_controller, settings_host_actions_semantics, settings_search_index, winui_settings_navigation, winui_dock_page_presenter, settings_update_rules; 02-tests.log.
- 主程序输入未变，引用 f6fb9750 的标准 Release 构建；本提交只编译受影响测试 / Host inputs unchanged; reuse f6fb9750 standard Release build; this commit compiled affected tests only.
- 桌面实机与全量测试仍待后续 / Desktop runtime acceptance and full tests remain pending.

+++ tests/dock_and_window_rules_tests.cpp
@@ -953,7 +953,7 @@ int main(int argc, char** argv)
             "status bar must use the final Shell-approved rectangle");
         Check(reservation.Registered(), "hiding a registered bar must not require removing its reservation");
         reservation.Remove(); reservation.Remove();
-        Check(std::count(messages.begin(), messages.end(), ABM_REMOVE) == 1,
+        Check(std::count(messages.begin(), messages.end(), static_cast<DWORD>(ABM_REMOVE)) == 1,
             "closing an AppBar releases its reservation exactly once");
         messages.clear();
         snowdesktop::StatusBarAppBar rejected([&](DWORD message, APPBARDATA&) -> UINT_PTR {
